refactor: prune packages and refine UI export abstraction
