+++ apps/api/src/index.ts
@@ -1,6 +1,5 @@
 import { Hono } from "hono";
 import { drizzle } from "drizzle-orm/d1";
-import { auth as createAuth } from "@ously/auth";
 import * as schema from "@ously/db";
 
 type Bindings = {
@@ -13,18 +12,13 @@ app.get("/", (c) => {
   return c.text("Ously API - Online");
 });
 
-app.get("/me", async (c) => {
+/**
+ * Example route using the shared DB package
+ */
+app.get("/users", async (c) => {
   const db = drizzle(c.env.DB, { schema });
-  const auth = createAuth(db);
-  const session = await auth.api.getSession({
-    headers: c.req.raw.headers,
-  });
-
-  if (!session) {
-    return c.json({ error: "Unauthorized" }, 401);
-  }
-
-  return c.json({ user: session.user });
+  const users = await db.select().from(schema.users).all();
+  return c.json({ users });
 });
 
 export default app;
