--- packages/auth/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@ously/auth",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "dependencies": {
-    "@ously/db": "workspace:*",
-    "better-auth": "^1.1.21"
-  },
-  "devDependencies": {
-    "@ously/tsconfig": "workspace:*",
-    "typescript": "^5.4.5"
-  }
-}
