--- packages/auth/src/index.ts
@@ -1,18 +1,0 @@
-import { betterAuth } from "better-auth";
-import { drizzleAdapter } from "better-auth/adapters/drizzle";
-import * as schema from "@ously/db";
-
-export const auth = (db: any) => betterAuth({
-    database: drizzleAdapter(db, {
-        provider: "sqlite",
-        schema: {
-            user: schema.users,
-            session: schema.sessions,
-            account: schema.accounts,
-            verification: schema.verifications,
-        }
-    }),
-    emailAndPassword: {
-        enabled: true
-    }
-});
