--- packages/auth/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@ously/tsconfig/base.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src"
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"]
-}
