--- packages/payments/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@ously/payments",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "dependencies": {
-    "@polar-sh/sdk": "^0.8.1"
-  },
-  "devDependencies": {
-    "@ously/tsconfig": "workspace:*",
-    "typescript": "^5.4.5"
-  }
-}
