+++ packages/ui/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./components/button";
-export * from "./lib/utils";
